Integrated the `slug` functionality of Hugo in to the posts.
I want the post files to be listed with a number as a prefix so that I
get the same listing in Finder as I do in the posts section in a
browser.

However, this approach puts that number prefix in to the URL as well which does
not look nice. Slugs allow me to modify the URL as I like.

It's a bit cumbersome to write the title twice but I believe it to be a worthy
endeavour all in all: good structure and agile handling.